+++ sdk-python/obsvr/events.py
@@ -28,6 +28,7 @@ _RESERVED_META_KEYS = (
     "obsvr_span",
     "obsvr_telemetry",
     "obsvr_external_backend",
+    "obsvr_integrity_flags",
 )